Add files via upload
This macro opens .lif or .nd2 files and analyzes them for each channel based on a threshold that you decide.

In order to use this, you will have to open your images in advance and determine what threshold for each channel is appropriate. Make a note of these values as you will need to enter them into the macro for it to run.